==> Step 2 - Combine Files.ipynb <==
import os
import pandas as pd
import joblib
from tqdm import tqdm
import datetime
from datetime import datetime
from bs4 import BeautifulSoup

%%time

a = pd.DataFrame({'filename1': os.listdir()})
b = a.filename1.str.startswith("security")
files = list(a[b].filename1)
final_df = pd.DataFrame()

for file in tqdm(files):
    df = joblib.load(file)
    if final_df.empty:
        final_df = df
    else:
        final_df = pd.concat([df, final_df])

final_df.sample(2)

sum(list(final_df.URL == final_df.link))

final_df2 = final_df[['title', 'summary_x', 'URL', 'published', 'keywords', 'summary_y', 'text' ]]

------

# the column 'published' has the time string.  Convert to datetime, as a timezone aware entry. 
# Fix EDT and PDT in advance

t1 = final_df2.published.apply(str)
t1 = [x.replace('EDT', '-0400') for x in t1]
t1 = [x.replace('PDT', '-0700') for x in t1]
t1 = [(pd.to_datetime(x)) for x in t1]
final_df2.published = t1

# Write a function to get just the data and time without timezone so it can be exported to Excel

def no_timezone(time_string):
    try:
        no_tz = datetime.strftime(time_string, "%Y-%m-%d %H:%M:%S")
    except:
        no_tz = 0
    return no_tz

# Get the published date (no timezone) in a new column
final_df2['published_date'] = final_df2.published.apply(no_timezone)

# Drop the published column 
final_df2.drop(columns = ['published'], inplace = True)

len(final_df2)

final_df2.sample(5)

# print((final_df2.applymap(type) == list).all())

# print(final_df2.applymap(type))

# a = ['one', 'two', 'three']
# b = ','.join(a)


final_df2.keywords = [','.join(a) for a in final_df2.keywords.copy(deep=True)]

print('Length before drop_duplicates', len(final_df2))
final_df2 = final_df2.drop_duplicates()
print('Length after drop_duplicates', len(final_df2))


final_df2.text.nunique()
# final_df2.text2=final_df2.text

final_df3 = final_df2.groupby('text').head(1)
print(len(final_df3), 'articles in final_df3')

final_df3.reset_index(drop=True, inplace=True)

final_df3


final_df3.sample(3)

# sum(final_df3.title.str.contains("Corona|corona|coronavirus|covid|Covid|cyber"))

final_df3.to_excel('temp.xlsx',index=False)

# Define a function for True/False for corona depending on if any of the corona related
# strings is found in a piece of text

def corona(text):
    corona_story_strings = ['covid', 'corona', 'sars', 'virus', 'coronavirus', 'vaccine']
    return any(x in text for x in corona_story_strings)
    
    


final_df3['corona'] = final_df3.text.apply(corona)

joblib.dump(final_df3, 'final_df3.pkl')

final_df3.shape

final_df3.corona.value_counts()

final_df3.query('corona == True')

final_df3.query('corona == True').to_excel('temp.xlsx')

!pip install tabloo
