==> Step 5a - FirstExtractiveSummarization.ipynb <==
### Extractive text summarization
This notebook has a function at the end that can be used on any 

from tqdm import tqdm
import joblib
data = joblib.load('final_df3.pkl')

def wrap(mytext):
  import textwrap
  wrapper = textwrap.TextWrapper(width=60)
  word_list = wrapper.wrap(text=mytext) 

  # Print each line. 
  for element in word_list: 
      print(element) 


print(data.columns)
textonly = list(data.text)

# Main function generate_summary defined in this cell, usage is generate_summary( text, number_of_sentences_to_return)
# https://towardsdatascience.com/understand-text-summarization-and-create-your-own-summarizer-in-python-b26a9f09fc70

from nltk.corpus import stopwords
from nltk.cluster.util import cosine_distance
import numpy as np
import networkx as nx
import nltk
nltk.download('stopwords')
import textwrap
 
def read_article(text):
  # file = open(file_name, 'r')
  # filedata = file.readlines()
  # filedata = [x for x in filedata if x != '\n']
  filedata = text
  filedata = [x.replace('\n',' ') for x in filedata]
  filedata = ''.join(filedata)
  filedata = filedata.split('. ')
  sentences = []

  for sentence in filedata:
    sentences.append(sentence.replace('[^a-zA-Z]', ' ').split(' '))
     
  return sentences

def sentence_similarity(sent1, sent2, stopwords=None):
    if stopwords is None:
        stopwords = []
 
    sent1 = [w.lower() for w in sent1]
    sent2 = [w.lower() for w in sent2]
 
    all_words = list(set(sent1 + sent2))
 
    vector1 = [0] * len(all_words)
    vector2 = [0] * len(all_words)
 
    # build the vector for the first sentence
    for w in sent1:
        if w in stopwords:
            continue
        vector1[all_words.index(w)] += 1
 
    # build the vector for the second sentence
    for w in sent2:
        if w in stopwords:
            continue
        vector2[all_words.index(w)] += 1
 
    return 1 - cosine_distance(vector1, vector2)
 
def build_similarity_matrix(sentences, stop_words):
    # Create an empty similarity matrix
    similarity_matrix = np.zeros((len(sentences), len(sentences)))
 
    for idx1 in range(len(sentences)):
        for idx2 in range(len(sentences)):
            if idx1 == idx2: #ignore if both are same sentences
                continue 
            similarity_matrix[idx1][idx2] = sentence_similarity(sentences[idx1], sentences[idx2], stop_words)

    return similarity_matrix


def generate_summary(file_name, top_n=5):
    stop_words = stopwords.words('english')
    summarize_text = []

    # Step 1 - Read text anc split it
    sentences =  read_article(file_name)
    if len(sentences) < top_n:
      top_n = len(sentences)

    # Step 2 - Generate Similary Martix across sentences
    sentence_similarity_martix = build_similarity_matrix(sentences, stop_words)

    # Step 3 - Rank sentences in similarity martix
    sentence_similarity_graph = nx.from_numpy_array(sentence_similarity_martix)
    scores = nx.pagerank(sentence_similarity_graph)

    # Step 4 - Sort the rank and pick top sentences
    ranked_sentence = sorted(((scores[i],s) for i,s in enumerate(sentences)), reverse=True)    
    #print("Indexes of top ranked_sentence order are ", ranked_sentence)    

    for i in range(top_n):
      summarize_text.append(" ".join(ranked_sentence[i][1]))

    # Step 5 - Output the summarized text
    summary = ". ".join(summarize_text)
    # print("Summarize Text: \n", ". ".join(summarize_text))
    return summary

    # wrapper = textwrap.TextWrapper(width=60)
    # word_list = wrapper.wrap(text=summary) 
  
    # # Print each line. 
    # for element in word_list: 
    #     print(element) 

    # # print(type(summary))
    # # print(summary)
    

# let's begin
# generate_summary( "msft.txt", 2)

summ = []
err = []
for x in tqdm(textonly):
  try:
    summ.append(generate_summary(x, 5))
  except:
    err.append(textonly.index(x))
    summ.append('Could not generate')
    

    


x = textonly[345]
textonly[345]

len(textonly[:3])

file = open('input.txt','w')
for element in textonly[:4]:
    file.write(element+'\n')
file.close()

generate_summary(x, 4)

err

data.iloc[795].URL

i=3445
print(wrap(summ[i]))
print('\n-------------')
print(wrap(summ[i]))

wrap(textonly[796])

wrap(summ[2])

data[1]

read_article(textonly)

import os
os.chdir("/content/drive/My Drive/Colab Notebooks")

    stop_words = stopwords.words('english')
    summarize_text = []

    # Step 1 - Read text anc split it
    sentences =  read_article('textfile.txt')

    # Step 2 - Generate Similary Martix across sentences
    sentence_similarity_martix = build_similarity_matrix(sentences, stop_words)

    # Step 3 - Rank sentences in similarity martix
    sentence_similarity_graph = nx.from_numpy_array(sentence_similarity_martix)
    scores = nx.pagerank(sentence_similarity_graph)

len(read_article(textonly[0]))
